deploy/huggingface/requirements.txt
llama-cpp-python==0.2.90
flask==3.0.0
gunicorn==21.2.0
huggingface-hub==0.20.0
